requirements.txt
python-dotenv==1.0.1
aiogram==3.17.0
asyncpg==0.30.0
singleton-decorator==1.0.0
aiosmtplib==4.0.0
